Extract components test utils
