+++ utils/components-test-utils/src/main/scala/pl/touk/nussknacker/engine/util/test/TestScenarioRuntime.scala
@@ -1,0 +1,14 @@
+package pl.touk.nussknacker.engine.util.test
+
+import com.typesafe.config.Config
+import pl.touk.nussknacker.engine.graph.EspProcess
+
+trait TestScenarioRuntime {
+  val config: Config
+
+  def run(scenario: EspProcess): Unit
+
+  def produceData(dataGenerator: () => Any): Unit = {}
+
+  def results(): Any = {}
+}
